pipelines/project7-phase6/requirements.txt
kfp==2.7.0
google-cloud-bigquery==3.25.0
